moved classes related to the Inventory system into their own packages and sub packages
