+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/inventory/cards/ICardDeconstructor.java
@@ -1,4 +1,4 @@
-package com.internetEnemies.combatCritters.Logic;
+package com.internetEnemies.combatCritters.Logic.inventory.cards;
 
 import com.internetEnemies.combatCritters.objects.Card;
