+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/inventory/decks/DeckBuilder.java
@@ -9,7 +9,7 @@
  *               therefore must be created by the DeckManager.getBuilder()
  */
 
-package com.internetEnemies.combatCritters.Logic;
+package com.internetEnemies.combatCritters.Logic.inventory.decks;
 
 
 import com.internetEnemies.combatCritters.objects.Card;
