+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/inventory/decks/DeckManager.java
@@ -8,7 +8,7 @@
  *               provide DeckBuilder class for each deck using getBuilder
  */
 
-package com.internetEnemies.combatCritters.Logic;
+package com.internetEnemies.combatCritters.Logic.inventory.decks;
 
 import com.internetEnemies.combatCritters.data.Database;
 import com.internetEnemies.combatCritters.data.ICardInventory;
