+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/inventory/decks/IDeckManager.java
@@ -7,7 +7,7 @@
  * @PURPOSE:    interface for managing multiple decks
  */
 
-package com.internetEnemies.combatCritters.Logic;
+package com.internetEnemies.combatCritters.Logic.inventory.decks;
 
 import com.internetEnemies.combatCritters.objects.DeckDetails;
