+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/inventory/decks/IDeckValidator.java
@@ -1,4 +1,4 @@
-package com.internetEnemies.combatCritters.Logic;
+package com.internetEnemies.combatCritters.Logic.inventory.decks;
 
 import com.internetEnemies.combatCritters.data.IDeck;
 import com.internetEnemies.combatCritters.objects.DeckValidity;
