+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/inventory/decks/IOnDeckChange.java
@@ -1,4 +1,4 @@
-package com.internetEnemies.combatCritters.Logic;
+package com.internetEnemies.combatCritters.Logic.inventory.decks;
 
 import com.internetEnemies.combatCritters.objects.DeckValidity;
