+++ app/src/main/java/com/internetEnemies/combatCritters/Logic/inventory/packs/PackBuilder.java
@@ -7,7 +7,7 @@
  * @PURPOSE:    Builder for pack
  */
 
-package com.internetEnemies.combatCritters.Logic;
+package com.internetEnemies.combatCritters.Logic.inventory.packs;
 
 import com.internetEnemies.combatCritters.objects.Card;
 import com.internetEnemies.combatCritters.objects.CardSlot;
