+++ app/src/main/java/com/internetEnemies/combatCritters/data/hsqldb/RegistryPackHSQLDB.java
@@ -1,6 +1,6 @@
 package com.internetEnemies.combatCritters.data.hsqldb;
 
-import com.internetEnemies.combatCritters.Logic.PackBuilder;
+import com.internetEnemies.combatCritters.Logic.inventory.packs.PackBuilder;
 import com.internetEnemies.combatCritters.data.IRegistry;
 import com.internetEnemies.combatCritters.data.hsqldb.DSOHelpers.PackHelper;
 import com.internetEnemies.combatCritters.objects.Card;
