+++ app/src/test/java/com/internetEnemies/combatCritters/IntegrationTests/CardCatalogTest.java
@@ -2,8 +2,8 @@ package com.internetEnemies.combatCritters.IntegrationTests;
 
 import static org.junit.Assert.assertEquals;
 
-import com.internetEnemies.combatCritters.Logic.CardCatalog;
-import com.internetEnemies.combatCritters.Logic.ICardCatalog;
+import com.internetEnemies.combatCritters.Logic.inventory.cards.CardCatalog;
+import com.internetEnemies.combatCritters.Logic.inventory.cards.ICardCatalog;
 import com.internetEnemies.combatCritters.TestUtils;
 import com.internetEnemies.combatCritters.data.ICardInventory;
 import com.internetEnemies.combatCritters.data.hsqldb.CardInventoryHSQLDB;
