+++ app/src/test/java/com/internetEnemies/combatCritters/IntegrationTests/CardDeconstructorIntegrationTest.java
@@ -12,8 +12,8 @@ import static junit.framework.TestCase.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
-import com.internetEnemies.combatCritters.Logic.CardDeconstructor;
-import com.internetEnemies.combatCritters.Logic.ICardDeconstructor;
+import com.internetEnemies.combatCritters.Logic.inventory.cards.CardDeconstructor;
+import com.internetEnemies.combatCritters.Logic.inventory.cards.ICardDeconstructor;
 import com.internetEnemies.combatCritters.TestUtils;
 import com.internetEnemies.combatCritters.data.ICardInventory;
 import com.internetEnemies.combatCritters.data.ICurrencyInventory;
